Extend allowed DataSources
This PR extends the supported DataSources handled by the
external-provisioner to include PVCs (Clone) and external API Object
types.

It removes the restriction of of the provisioner erroring when a
DataSource.Kind of anything other than VolumeSnapshot is supplied
and adds handling for a DataSource.Kind of PersistentVolumeClaim.

In the case of specified DataSources that the provisioner doesn't deal
with, we just log that the provisioner doesn't take any action on the
specified type, and continue with provisioning ignoring the DataSource
field.

This not only enables Clone requests, but it also allows development of
external controllers to initialize newly created volumes.  Additional
improvements/enhancements will be added via PV/PVC taints and
tolerations when they're implemented.

Addresses: Issue #172