better deprecated flag handling + vendor update
The warning for the deprecated connection-timeout was only printed
when it was set to a non-default value. When set explicitly to
zero (admittedly unlikely), no warning was printed.

This approach also gets rid of the actual variable and thus catches
code which accidentally still uses it (see
https://github.com/kubernetes-csi/external-provisioner/pull/239).

This depends on a more recent csi-lib-utils. Because the previous
vendoring was done with an older dep version, "dep ensure -update"
cannot update just that project, so everything gets updated instead.