replace "dep" with "go mod", update to Kubernetes 1.15.4
A simple "go mod init && go mod tidy" picked incompatible revisions of
the k/k modules, so go-get-kubernetes.sh had to be used which in turn
implied that Kubernetes had to be updated, because a release with
go.mod file is needed by that script. It makes sense to update anyway.

However, there's one exception: the "replace" for
k8s.io/csi-translation-lib had to be removed manually and then a
compatible revision of it was picked with "go get
k8s.io/csi-translation-lib@17db17aaadeb70b1d7259c18ff045f51a1164d39"
(same as in external-attacher).

The external-provisioner code would have to be modified to work with
the newer lib.