update to client-go with ctx parameter
Kubernetes 1.18 introduced a ctx parameter in all API client-go API
calls. Whenever possible, that context is passed in by the caller
instead of making one up locally.

Newer releases of external-snapshotter and
sig-storage-lib-external-provisioner are needed because they have to
be compatible with the new client-go API (can't mix different versions
of it).