capacity: walk up ownership chain
Generic code from the controller runtime is used to retrieve
unstructured objects. This is needed for deployment via Deployment,
because the direct parent of the pod is then a ReplicaSet which itself
will get deleted by the Deployment when rolling out changes.

There are intentionally no unit tests for the feature because that
would bring in even more additional dependencies.