storage capacity: optional ownership, use labels instead
The mandatory ownership reference prevented deploying
external-provisioner outside of the Kubernetes cluster and made it
impossible to reuse objects when rolling out an updated CSI driver
which used distributed provisioning.

It's better to identify managed objects with labels. This also makes
it possible to optimize the watch (to be implemented in a separate
commit).