capacity test: create CSIStorageObject before instantiating client-set
This is necessary to ensure that the controller sees all of the
objects when it starts.

The previous approach with Create() was used to ensure that namespace
and UID are set. We now have to take care of that manually.

(cherry picked from commit 571be5a9e5b1dcddcfc1895048ef3aa17bab5aea)