capacity: support both v1 and v1beta1
The code itself uses the v1 API and directly uses the normal client-go API when
v1 is supported by the server. If the server doesn't support that API, wrappers
around the v1beta1 API convert CSIStorageCapacity objects back and forth as
needed.

The reason for this more complex solution is that it avoids a breaking change
in the external-provisioner and thus provides a smoother transition path: CSI
driver developers can update to the next external-provisioner release and use
it both with Kubernetes < 1.24 and >= 1.24.