capacity: use separate client and lower QPS/Burst settings
The rationale behind this change is that it allows external-provisioner to deal
with volumes without being slowed down by the concurrent CSIStorageCapacity
updates. Those updates are less important and the scheduler can deal with
outdated information, so enforcing a lower rate by default is acceptable.

If few changes are needed, updating CSIStorageCapacity keeps up with volume
changes. If many volumes get created or deleted quickly, throttling the
updating is beneficial because multiple changes can be combined into a single
update.

This has helped in practice during a scale test with 100 nodes and 10000
pods+volumes. Without this change, there were intermittent problems with an
unresponsive apiserver. With this change, no such problems occurred.