Seed random generator before creating provisioner identity
The Go math/random package needs to be seeded otherwise it will generate the same sequence over and over.

Without seeding, all provisioners will have `8081` as part of their random ID, as that is the first number generated by the default sequence.

Use the millisecond timeStamp created just above as seed. This still links the random number to the time stamp, but that's probably good enough for its purpose.