cleanup: build the test case objects with all the information
Remove the conditional statements that customize the input test
case object based on the different test scenarios in the function
of running a test.

The function to build the test cases should build the test case
objects with all the information that the function of running a
test needs so that it makes the test code easier to read and
understand.

Signed-off-by: Hidehisa Shitomi <hidehisa.shtm@gmail.com>