Set ownerRef of CSIStorageCapacity to node when using node deployment
This change adds logic to set the OwnerReference of the
CSIStorageCapacity to the node itself when using node deployment
as in this case the external-provisioner will be deployed as a
DaemonSet together with the CSI plugin for provisioning node
local volumes. Under this case, we should set the OwnerReference
of the corresponding CSIStorageCapacity to the node itself so
that when the node is gone (e.g. removed by cluster-autoscaler),
the CSIStorageCapacity will be garbage-collected and not left
behind.